requirements.txt
django
simple-salesforce
hubspot-api-client
requests
python-dotenv